pyproject.toml
[project]
name = "network-optimization"
version = "0.1.0"
description = "Optimize network parameters using Q-Learning reinforcement learning on cell KPI states"
authors = [
    { name = "Adityo Nugroho", email = "adityo.nugroho.id@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    # Core ML/Data Science
    # NOTE: Version constraints for SHAP compatibility
    "pandas>=2.1.0",
    "numpy>=1.24.0,<2.0",      # SHAP has color conversion bug with NumPy 2.x
    "scikit-learn>=1.3.0",

    # ML Models
    # NOTE: XGBoost pinned for SHAP compatibility
    "xgboost>=1.7.0,<2.0",     # SHAP not compatible with XGBoost 2.x base_score format
    "lightgbm>=4.1.0",

    # Model Interpretability
    "shap>=0.42.0",
    "numba>=0.59.0",           # Required for Python 3.12 support

    # Visualization
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",

    # Jupyter
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",

    # Utilities
    "tqdm>=4.66.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "N803",  # argument name should be lowercase (ML convention: X, X_train)
    "N806",  # variable name should be lowercase (ML convention: X, X_train)
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
